Text Translator/requirements.txt
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
libretranslatepy==2.1.1
lxml==4.6.3
Pillow==8.3.2
requests==2.26.0
rfc3986==1.5.0
sniffio==1.2.0
tk==0.1.0
translate==3.6.1
urllib3==1.26.7
